external/ipw_internal/intelligence-per-watt/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "intelligence-per-watt"
version = "0.1.0"
description = "Intelligence Per Watt - Energy-aware LLM inference profiling and orchestration"
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    # Core ipw dependencies
    "httpx>=0.27",
    "aiohttp>=3.12.15",
    "rich>=13",
    "click>=8",
    "matplotlib>=3.9",
    "grpcio>=1.62",
    "protobuf>=5",
    "pytest>=8",
    "requests>=2.31.0",
    "datasets>=2.20",
    "numpy>=1.26",
    "pyarrow>=15",
    "tqdm>=4.66",
    "hf_transfer",
]

[project.optional-dependencies]
# Inference client backends
ollama = ["ollama>=0.1.0"]
vllm = ["vllm>=0.11.0"]

# Agents package dependencies
agents = [
    "fastapi>=0.121.3",
    "uvicorn>=0.38.0",
    "agno>=0.1.0",
    "pydantic>=2.0.0",
    "terminal-bench>=0.2.18",
    "openhands>=1.7.1",
    "tavily-python>=0.3.0",  # For agents.mcp.WebSearchServer
]

# Evals package dependencies
evals = [
    "loguru>=0.7.3",
    "docker>=7.0.0",
    "huggingface-hub>=0.19.0",
    "pandas>=2.0.0",
    "litellm>=1.80.11",
    "python-dotenv>=1.0.0",
    "tau2",
    "swebench>=3.0.0",
    "sweagent",
    "swe-rex>=1.4.0",
    "nest-asyncio>=1.6.0",
]

# Orchestrator training dependencies
orchestrator = [
    "trl>=0.8.0",
    "accelerate>=0.27.0",
    "transformers>=4.38.0",
    "torch>=2.8.0",
    "openai>=1.12.0",
    "anthropic>=0.20.0",
    "pyyaml>=6.0",
    "dotenv>=0.9.9",
    "google-genai>=1.57.0",
    "tavily-python>=0.7.19",
    "verl>=0.7.0"
]

# GAIA benchmark support
gaia = [
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "python-pptx>=0.6.0",
    "Pillow>=10.0.0",
    "mutagen>=1.47.0",
    "openpyxl>=3.1.0",
]

# BrowseComp benchmark support
browsecomp = [
    "trafilatura>=1.6.0",
    "beautifulsoup4>=4.12.0",
    "tavily-python>=0.3.0",
]

# FinanceBench uses HuggingFace datasets
# No additional dependencies required
financebench = []

# Retrieval capabilities for benchmarks (FRAMES, BrowseComp, DeepResearch, etc.)
retrieval = [
    "rank-bm25>=0.2.2",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.0",
    "tiktoken>=0.5.0",
]

# Grid eval analysis
grid = [
    "tabulate>=0.9",
]

# TUI for orchestrator
tui = [
    "textual>=0.47.0",
]

# Dataset generator dependencies
dataset-generator = [
    "torch>=2.8.0",
    "huggingface-hub>=0.19.0",
    "datasets>=2.0.0",
]

# Inference simulator dependencies
inference-simulator = [
    "scikit-learn>=1.3.0",  # For RandomForestRegressor and other sklearn estimators
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "scipy>=1.10.0",
    "ppi-python>=0.2.0",
]

# Inference search dependencies (uses inference_simulator + ipw.simulator)
inference-search = [
    "bayesian-optimization>=1.4.0",
]

# All optional dependencies (excludes orchestrator due to verl numpy<2 conflict on Python 3.13)
all = [
    "intelligence-per-watt[ollama,vllm,agents,evals,gaia,browsecomp,retrieval,grid,tui,dataset-generator,inference-simulator,inference-search]",
]

[project.scripts]
ipw = "ipw.cli:main"
serve-agent = "agents.serve:main"
dataset-generator = "dataset_generator.cli:main"
inference-search = "inference_search.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ipw", "src/agents", "src/evals", "src/orchestrator", "grid_eval", "src/dataset_generator", "src/inference_simulator", "src/inference_search"]

[tool.hatch.build.targets.wheel.force-include]
"src/ipw/telemetry/bin" = "ipw/telemetry/bin"

[tool.uv.sources]
terminal-bench = { git = "https://github.com/laude-institute/terminal-bench.git" }
tau2 = { git = "https://github.com/sierra-research/tau2-bench.git", rev = "337326e62d8e0ca74c353b004a9c5d748e0ba914" }
sweagent = { git = "https://github.com/SWE-agent/SWE-agent.git", rev = "0f3acafacabc0def8cc76b4e48acb4b6cf302cb9" }

[tool.pytest.ini_options]
markers = [
    "integration: marks tests as integration tests (may require hardware and running vLLM server)",
    "auto_server: marks tests that use the auto-server functionality",
    "submodel: marks tests that use multiple models/submodels",
]
timeout = 600
